memory_builder/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose
passlib
python-multipart
boto3
celery
redis
pydantic-settings
pydantic
python-dotenv
email-validator
python-json-logger
aiohttp
minio
gunicorn
azure-ai-inference